refactor tilecount
