+++ src/schema/hand.py
@@ -6,7 +6,8 @@ from pydantic import BaseModel
 
 from src.enum.common import CallType
 from src.schema.call import Call
-from src.schema.tile import Tile, TileCount
+from src.schema.count import TileCount
+from src.schema.tile import Tile
 
 
 class Hand(BaseModel):
