+++ src/schema/tile.py
@@ -1,5 +1,3 @@
-from typing import Union
-
 from pydantic import BaseModel
 
 from src.enum.common import TileType
@@ -21,31 +19,6 @@ class Tile(BaseModel):
             TileType.DRAGON: 30,
         }
         return start_idx[self.type] + self.value
-
-
-class TileCount:
-    def __init__(self, tiles: list[Tile]):
-        self._values = [0] * 34
-        for tile in tiles:
-            self._values[tile] += 1
-
-    def __getitem__(self, key: Union[Tile, int]):
-        return self._values[key]
-
-    def __setitem__(self, key: Union[Tile, int], value: int):
-        if value not in range(0, 5):
-            raise ValueError
-        self._values[key] = value
-
-    def __iter__(self):
-        yield from self._values
-
-    def get_last_nonzero_idx(self, idx: int = 0):
-        while idx < len(self._values):
-            if self._values[idx] != 0:
-                break
-            idx += 1
-        return idx
 
 
 class Tiles:
